pyproject.toml
[tool.poetry]
name = "acastorage"
version = "0.1.0"
description = ""
authors = ["Nina Jensen <jnik@aarhus.dk>"]
maintainers = ["Aarhus Stadsarkiv <stadsarkiv@aarhus.dk>"]
license = "GPL-3.0"
readme = "README.md"
homepage = "https://www.aarhusstadsarkiv.dk/"

[tool.poetry.dependencies]
python = "^3.8"
azure-storage-blob = "^12.7.0"
aiohttp = "^3.7.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^2.11"
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.800"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.14.0"
azure-identity = "^1.5.0"
azure-keyvault-secrets = "^4.2.0"

[tool.black]
line-length = 79
exclude = ".venv"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

